# Conflicts:
#	src/main/java/FileSystem/User.java
#	target/FileSystemWeb-1.0/WEB-INF/classes/FileSystem/Archive.class
#	target/FileSystemWeb-1.0/WEB-INF/classes/FileSystem/FileAdministrator.class
#	target/FileSystemWeb-1.0/WEB-INF/classes/FileSystem/Main.class
#	target/FileSystemWeb-1.0/WEB-INF/classes/servletPackage/ProcessCommand.class
#	target/classes/FileSystem/Archive.class
#	target/classes/FileSystem/FileAdministrator.class
#	target/classes/FileSystem/Main.class
#	target/classes/servletPackage/ProcessCommand.class

+++ src/main/java/FileSystem/FileAdministrator.java
@@ -21,6 +21,9 @@ import java.util.Date;
 public class FileAdministrator {
     
     public String moveArchive(String nameArchive, Folder folToMove, Folder folToDelete){
+        if(folToMove == null || folToDelete == null){
+            return "No se encontro un folder";
+        }
         Archive archive = folToDelete.getArchive(nameArchive);
         if(archive != null){
             if(folToDelete.deleteArchive(archive)){
@@ -34,6 +37,9 @@ public class FileAdministrator {
     }
     
     public String moveFolder(Folder folder, Folder folToMove){
+        if(folToMove == null || folder == null){
+            return "No se encontro un folder";
+        }
         if(!folToMove.verNameFolder(folder.getName())){
             if(folder.getFather().deleteFolder(folder)){
                 folToMove.addFolder(folder);
@@ -46,6 +52,9 @@ public class FileAdministrator {
     }
     
     public String createFolder(Folder folder, String name, String directory, String createDate, String user, String locationLogic){
+        if(folder == null){
+            return "No se encontro un folder";
+        }
         if(!folder.verNameFolder(name)){
             Folder createFolder = new Folder(name, directory, folder, createDate, user, locationLogic);
             folder.addFolder(createFolder);
@@ -54,15 +63,22 @@ public class FileAdministrator {
         return "No se logro crear correctamente el folder";
     }
 
-    public String createFile(Folder father, String name, String extension, String fileContent, String logicalLocation) {
+
+    public String createFile(Folder father, String name, String extension, String fileContent, String logicalLocation, User user) {
+        if(father == null){
+            return "No se encontro un folder";
+        }
         if(!father.verNameArchive(name)) {
             String currentDate = getCurrentDate();
             int size = fileContent.getBytes().length;
-            Archive archive = new Archive(name, size, extension, currentDate, currentDate, 666, 0, fileContent, logicalLocation);
-            father.addArchive(archive);
-            return "Se creó el archivo correctamente";
+            boolean enoughSize = user.calculateSize(size, 0);
+            if (enoughSize) {
+                Archive archive = new Archive(name, size, extension, currentDate, currentDate, 666, 0, fileContent, logicalLocation);
+                father.addArchive(archive);
+                return "Se creó el archivo correctamente";
+            }
         }
-        return "No se pudo crear el archivo";
+        return "No se pudo crear el archivo";
     }
     
     public String getCurrentDate() {
@@ -72,20 +88,37 @@ public class FileAdministrator {
         return currentDate;
     }
 
-    public String updateFile(Folder folder, String fileName, String fileContent) {
-        if(folder.verNameArchive(fileName)) {
-            String currentDate = getCurrentDate();
-            int size = fileContent.getBytes().length;
-            Archive archive = folder.getArchive(fileName);
-            archive.setFileContent(fileContent);
-            archive.setSize(size);
-            archive.setDateModify(currentDate);
-            return "Se actualizó el archivo correctamente";
-        }
-        return "No se pudo actualizar el archivo";
+
+    public String updateFile(Folder folder, String fileName, String fileContent, User user,MainFileSystem fs) {
+            if(folder.verNameArchive(fileName)) {
+                String currentDate = getCurrentDate();
+                int size = fileContent.getBytes().length;
+                Archive archive = folder.getArchive(fileName);
+                boolean enoughSize = user.calculateSize(size, archive.getSize());
+                if (enoughSize) {
+                    archive.setFileContent(fileContent);
+                    archive.setSize(size);
+                    archive.setDateModify(currentDate);
+                    for(int i = 0; i < archive.sharedUsers.size(); i++){
+                        Folder shared = fs.getU(archive.sharedUsers.get(i)).sharedFolder;
+                        for(int j= 0; j < shared.archiveIn.size();j++){
+                            if(shared.archiveIn.get(j).getName().equals(fileName)){
+                                shared.archiveIn.remove(j);
+                                shared.archiveIn.add(archive);
+                            }
+                        }
+                    }
+                    return "Se actualizó el archivo correctamente";
+                }
+                
+            }
+            return "No se pudo actualizar el archivo";
     }
 
     public String seeFileProperties(Folder folder, String fileName) {
+        if(folder == null){
+            return "No se encontro un folder";
+        }
         String properties = "";
         if(folder.verNameArchive(fileName)) {
             Archive archive = folder.getArchive(fileName);
@@ -99,6 +132,9 @@ public class FileAdministrator {
     }
 
     public String seeFile(Folder folder, String fileName) {
+        if(folder == null){
+            return "No se encontro un folder";
+        }
         String fileContent = "";
         if(folder.verNameArchive(fileName)) {
             Archive archive = folder.getArchive(fileName);
@@ -108,6 +144,9 @@ public class FileAdministrator {
     }
     
     public String listarDirectorios(Folder folder){
+        if(folder == null){
+            return "No se encontro un folder";
+        }
         String lista = "";
         ArrayList<Folder> folders = folder.getFoldersIn();
         for(int i = 0; i < folders.size(); i++){
@@ -122,16 +161,31 @@ public class FileAdministrator {
         return lista;
     }
     
-    public String deleteFile(Folder folder, String fileName) {
+
+    public String deleteFile(Folder folder, String fileName, MainFileSystem fs) {
+        if(folder == null){
+            return "No se encontro un folder";
+        }
         boolean deleted = false;
         if(folder.verNameArchive(fileName)) {
             Archive archive = folder.getArchive(fileName);
             deleted = folder.deleteArchive(archive);
+            for(int i = 0; i < archive.sharedUsers.size(); i++){
+                Folder shared = fs.getU(archive.sharedUsers.get(i)).sharedFolder;
+                for(int j= 0; j < shared.archiveIn.size();j++){
+                    if(shared.archiveIn.get(j).getName().equals(fileName)){
+                        shared.archiveIn.remove(j);
+                    }
+                }
+            }
         }
         return deleted ? "Se eliminó el archivo correctamente" : "No se pudo eliminar el archivo";
     }
 
     public String deleteFolder(Folder folder, String folderName) {
+        if(folder == null){
+            return "No se encontro un folder";
+        }
         boolean deleted = false;
         if(folder.verNameFolder(folderName)) {
             Folder folderToDelete = folder.getFolder(folderName);
@@ -141,6 +195,9 @@ public class FileAdministrator {
     }   
     
     public String copiarVVFolder(Folder folder, Folder folToCopy){
+        if(folder == null || folToCopy == null){
+            return "No se encontro un folder";
+        }
         if(folToCopy.verNameFolder(folder.getName())){
             Folder newFolder = new Folder(folder.getName(), folToCopy.getDirectory(), 
                     folToCopy, getCurrentDate(), folder.getUser(), 
@@ -152,7 +209,11 @@ public class FileAdministrator {
         }
         return "No se puede copiar el folder.";
     }
+    
     public String copiarVVArchive(Folder folder, String nameArchive, Folder currentFolder){
+        if(folder == null || currentFolder == null){
+            return "No se encontro un folder";
+        }
         Archive archive = currentFolder.getArchive(nameArchive);
         if(folder.verNameArchive(archive.getName())){
             Archive newArchive = new Archive(archive.getName(), archive.getSize(), 
@@ -163,6 +224,7 @@ public class FileAdministrator {
         }
         return "No se puede copiar el archivo.";
     }
+    
     public String changeDirectory(User u, String path){
         if(path.equals("..")){
             u.currentFolder = u.currentFolder.getFather();
@@ -199,6 +261,9 @@ public class FileAdministrator {
     }
         
     public String vrArchive(String fileName, Folder folder){
+        if(folder == null){
+            return "No se encontro un folder";
+        }
         if(folder.verNameArchive(fileName)) {
             Archive archive = folder.getArchive(fileName);
             try (BufferedWriter writer = new BufferedWriter(new FileWriter(fileName+".txt"))) {
@@ -207,6 +272,14 @@ public class FileAdministrator {
             } catch (IOException e) {
                 return "No se descargar correctamente el archivo";
             }
+        }
+        return "No se logro encontrar el archivo";
+    }
+    public String shareFile(Folder folder, String fileName,User u){
+        if(folder.verNameArchive(fileName)) {
+            Archive archive = folder.getArchive(fileName);
+            u.sharedFolder.archiveIn.add(archive);
+            archive.sharedUsers.add(u.name);
         }
         return "No se logro encontrar el archivo";
     }
